backend update v2

+++ app/models/membership.model.js
@@ -7,11 +7,7 @@ module.exports = (sequelize, Sequelize) => {
           autoIncrement: true,
           primaryKey: true
         },
-        //id: {
-        //  type: Sequelize.STRING,
-        //  primaryKey: true,
-        //  defaultValue: Sequelize.UUIDV4, // Auto-generate UUIDs
-      //  },
+       
         title: {
           type: Sequelize.STRING,
           allowNull: false,
@@ -26,7 +22,7 @@ module.exports = (sequelize, Sequelize) => {
           defaultValue:'basic'
         },
         price: {
-          type: Sequelize.FLOAT,
+          type: Sequelize.DOUBLE,
           allowNull: false,
         },
       },
